fix: ajustes de erros do -Wall

+++ Grafo.cpp
@@ -114,8 +114,8 @@ bool Grafo::caminho_restrito(int v, int w, int x, int z, int marcado[]) {
 }
 
 bool Grafo::eh_clique(vector<int> vertices) {
-    for (int i = 0; i < vertices.size(); i++) {
-        for (int j = i + 1; j < vertices.size(); j++) {
+    for (size_t i = 0; i < vertices.size(); i++) {
+        for (size_t j = i + 1; j < vertices.size(); j++) {
             if (find(lista_adj_[vertices[i]].begin(), lista_adj_[vertices[i]].end(), vertices[j]) == lista_adj_[vertices[i]].end()) {
                 return false;
             }
